Rubics Cube is the best
Date: 08-11-2024

I love solving a rubics cube myself. But what I love even more, is making a program that will do it for me. I have here two python programs. One is solving using the blindfold method,  and the other is optimal. The last one is the same as the first, but in C++.